Split graph-degradation.js into health-plasma / deck-perturbation / wiring (#166)
graph-degradation.js mixed the WebGL health-plasma shader, the Cytoscape
deck-perturbation particle system + glitch factories, the rAF loop, and module
state in one 525-LOC file. Split into three:

- health-plasma.js — WebGL shader/program, canvas init+resize, heartbeat
  helpers, and a self-contained flow clock; exposes init/draw/stop. Renders
  only when WebGL is present and the overlay is visible; the clock advances
  every frame either way (preserving the original loop semantics).
- deck-perturbation.js — particle pool, all glitch factories, the
  severity-scaled spawner/update loop, and restore/discard; owns its
  containerEl + basePos + grid-glitch state.
- graph-degradation.js — thin wiring: owns the shared rAF loop, the params
  pulled from game state, and the #cy haze filter; drives both effects each
  frame.

Behavior-preserving. The public API (initGraphDegradation / updateFromState /
stopGraphDegradation) is unchanged, so visual-renderer + the preview harness
keep driving both effects through the same entry points. make check green.
